Add double before div int and div long

+++ src/ru/itmo/java/basics/lab2/Calculator.java
@@ -5,6 +5,7 @@ public class Calculator {
     public int add(int num1, int num2) {
         return (num1 + num2);
     }
+
     public double add(double num1, double num2) {
         return (num1 + num2);
     }
@@ -16,6 +17,7 @@ public class Calculator {
     public int sub(int num1, int num2) {
         return (num1 - num2);
     }
+
     public double sub(double num1, double num2) {
         return (num1 - num2);
     }
@@ -27,6 +29,7 @@ public class Calculator {
     public int mult(int num1, int num2) {
         return (num1 * num2);
     }
+
     public double mult(double num1, double num2) {
         return (num1 * num2);
     }
@@ -35,15 +38,16 @@ public class Calculator {
         return (num1 * num2);
     }
 
-    public int div(int num1, int num2) {
-        return (num1 / num2);
+    public double div(int num1, int num2) {
+        return ((double) num1 / num2);
     }
+
     public double div(double num1, double num2) {
         return (num1 / num2);
     }
 
-    public long div(long num1, long num2) {
-        return (num1 / num2);
+    public double div(long num1, long num2) {
+        return ((double) num1 / num2);
     }
